+++ extensions/github/src/unread-notifications.tsx
@@ -36,6 +36,22 @@ function UnreadNotifications() {
     return response.data;
   });
 
+  const { data: iconsData } = useCachedPromise(
+    async (notifications) => {
+      if (!notifications) {
+        return [];
+      }
+      const icons = await Promise.all(
+        notifications.map((notification: Notification) => getNotificationIcon(notification)),
+      );
+      return icons;
+    },
+    [data],
+    {
+      keepPreviousData: true,
+    },
+  );
+
   const hasUnread = data && data.length > 0;
 
   async function markAllNotificationsAsRead() {
@@ -103,26 +119,27 @@ function UnreadNotifications() {
 
       <MenuBarExtra.Section>
         {hasUnread ? (
-          data.map((notification) => {
-            const icon = {
-              source: getNotificationIcon(notification).value,
-              tintColor: Color.PrimaryText,
-            };
+          data.map((notification, index) => {
+            const icon = iconsData?.[index];
             const title = notification.subject.title;
             const updatedAt = new Date(notification.updated_at);
             const tooltip = getNotificationTooltip(updatedAt);
 
+            if (!icon) {
+              return null;
+            }
+
             return (
               <MenuBarExtra.Item
                 key={notification.id}
-                icon={icon}
+                icon={{ source: icon.value["source"], tintColor: Color.PrimaryText }}
                 title={title}
                 subtitle={getNotificationSubtitle(notification)}
                 tooltip={tooltip}
                 onAction={() => openNotification(notification)}
                 alternate={
                   <MenuBarExtra.Item
-                    icon={icon}
+                    icon={{ source: icon.value["source"], tintColor: Color.PrimaryText }}
                     title={title}
                     subtitle="Mark as Read"
                     tooltip={tooltip}