+++ extensions/github/CHANGELOG.md
@@ -1,5 +1,10 @@
 # GitHub Changelog
 
+## [Fix icons for Notifications and Unread Notifications commands] - 2024-04-05
+
+- Fixes icons for issues and pull requests in `Notifications` and `Unread Notifications` commands.
+- Standardize the color for all the accessories in `My Pull Requests`.
+
 ## [Fix icon] - 2024-04-02
 
 - Fixes icon for closed pull requests.